Add decorator to handle warnings
