+++ swarmcg/shared/__init__.py
@@ -1,0 +1,1 @@
+from .utils import catch_warnings
